[NCCL] Add more details for checkForNCCLErrors
https://github.com/pytorch/pytorch/pull/45950 enhanced our NCCL logging errors so that we add some basic debug information about what when wrong when erroring out with a NCCL error.

However, that PR only used the added function for `C10D_NCCL_CHECK` which is used to check the return values of NCCL calls. However, in ProcessGroupNCCL we also have `checkForNCCLErrors` which checks for errors on nccl communicators, and in case of errors it would be good to have this logging there too.

Also renames the function s/errorMessage/getNcclErrorDetailStr

Differential Revision: [D27100497](https://our.internmc.facebook.com/intern/diff/D27100497/)

ghstack-source-id: 124090094
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54117